+++ test/storybook-playwright/storybook/main.js
@@ -7,7 +7,7 @@ const config = {
 	...baseConfig,
 	addons: [ '@storybook/addon-toolbars' ],
 	stories: [
-		'../../../packages/components/src/**/e2e.stories.@(js|tsx|mdx)',
+		'../../../packages/components/src/**/stories/e2e/*.@(js|tsx|mdx)',
 	],
 };
